requirements.txt
pandas==2.1.0
numpy==1.25.0
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.12.2
pdfplumber==0.10.0
PyMuPDF==1.23.1
pytesseract==0.3.10
beautifulsoup4==4.12.2
tensorflow==2.13.0
torch==2.1.0
torchvision==0.16.0
transformers==4.38.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
spacy==3.6.0
joblib==1.3.2